Fix sign and verify issues.
Currently the output of sign is an ArrayBuffer. However, this causes
issues when integrated with DataIntegrityProof, leading to an empty
signature value. Converting to Buffer fixes this issue.
Additionally, there are times when exportKeyPair is called with a
true value for secretKey, but no available secret key. To handle
this case, we need to also check if the key pair has a secretKey.